Move document loaders to a seperate file
